+++ app/src/main/java/entity/User.java
@@ -6,18 +6,18 @@ import java.io.Serializable;
  * 用户的实体类.
  * */
 public class User implements Serializable {
-    private int id;
+    private long id;
     private String nickName;
 
     public User() {
     }
 
-    public User(int id, String nickName) {
+    public User(long id, String nickName) {
         this.id = id;
         this.nickName = nickName;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
